+++ apps/ledger-live-desktop/src/newArch/features/ModularDrawer/screens/NetworkSelection/index.tsx
@@ -1,16 +1,19 @@
 import React from "react";
 import { CryptoOrTokenCurrency } from "@ledgerhq/types-cryptoassets";
-import { SelectNetwork as NetworksList } from "./components/List";
+import { SelectNetwork as NetworksList } from "./components";
 import { EnhancedModularDrawerConfiguration } from "@ledgerhq/live-common/wallet-api/ModularDrawer/types";
 import TrackDrawerScreen from "../../analytics/TrackDrawerScreen";
 import { MODULAR_DRAWER_PAGE_NAME } from "../../analytics/types";
+import { CurrenciesByProviderId } from "@ledgerhq/live-common/deposit/type";
 
 export type NetworkSelectionStepProps = {
   networks?: CryptoOrTokenCurrency[];
   networksConfiguration: EnhancedModularDrawerConfiguration["networks"];
   source: string;
   flow: string;
+  currenciesByProvider: CurrenciesByProviderId[];
   onNetworkSelected: (network: CryptoOrTokenCurrency) => void;
+  selectedAssetId?: string;
 };
 
 export function NetworkSelection({
@@ -19,6 +22,8 @@ export function NetworkSelection({
   source,
   flow,
   networksConfiguration,
+  currenciesByProvider,
+  selectedAssetId,
 }: Readonly<NetworkSelectionStepProps>) {
   return (
     <>
@@ -34,6 +39,9 @@ export function NetworkSelection({
         onNetworkSelected={onNetworkSelected}
         flow={flow}
         source={source}
+        networksConfig={networksConfiguration}
+        currenciesByProvider={currenciesByProvider}
+        selectedAssetId={selectedAssetId}
       />
     </>
   );