+++ apps/ledger-live-mobile/src/index.tsx
@@ -31,12 +31,14 @@ import {
   saveMarketState,
   saveTrustchainState,
   saveWalletExportState,
+  saveLargeMoverState,
 } from "./db";
 import {
   exportSelector as settingsExportSelector,
   osThemeSelector,
   hasSeenAnalyticsOptInPromptSelector,
   hasCompletedOnboardingSelector,
+  trackingEnabledSelector,
 } from "~/reducers/settings";
 import { accountsSelector, exportSelector as accountsExportSelector } from "~/reducers/accounts";
 import { exportSelector as bleSelector } from "~/reducers/ble";
@@ -94,6 +96,14 @@ import { useDeviceManagementKitEnabled } from "@ledgerhq/live-dmk-mobile";
 import { StoragePerformanceOverlay } from "./newArch/storage/screens/PerformanceMonitor";
 import { useDeviceManagementKit } from "@ledgerhq/live-dmk-mobile";
 import AppGeoBlocker from "LLM/features/AppGeoblocker";
+import { exportLargeMoverSelector } from "./reducers/largeMover";
+import {
+  TrackingConsent,
+  DatadogProvider,
+  AutoInstrumentationConfiguration,
+} from "@datadog/mobile-react-native";
+import { PartialInitializationConfiguration } from "@datadog/mobile-react-native/lib/typescript/DdSdkReactNativeConfiguration";
+import { initializeDatadogProvider } from "./datadog";
 
 if (Config.DISABLE_YELLOW_BOX) {
   LogBox.ignoreAllLogs();
@@ -123,12 +133,23 @@ function walletExportSelector(state: State) {
 function App() {
   const accounts = useSelector(accountsSelector);
   const analyticsFF = useFeature("llmAnalyticsOptInPrompt");
+  const datadogFF = useFeature("llmDatadog");
   const isLDMKEnabled = useDeviceManagementKitEnabled();
   const providerNumber = useEnv("FORCE_PROVIDER");
   const hasSeenAnalyticsOptInPrompt = useSelector(hasSeenAnalyticsOptInPromptSelector);
   const hasCompletedOnboarding = useSelector(hasCompletedOnboardingSelector);
   const dmk = useDeviceManagementKit();
   const dispatch = useDispatch();
+  const isTrackingEnabled = useSelector(trackingEnabledSelector);
+
+  const datadogAutoInstrumentation: AutoInstrumentationConfiguration = useMemo(
+    () => ({
+      trackErrors: datadogFF?.params?.trackErrors ?? false,
+      trackInteractions: datadogFF?.params?.trackInteractions ?? false,
+      trackResources: datadogFF?.params?.trackResources ?? false,
+    }),
+    [datadogFF?.params],
+  );
 
   useEffect(() => {
     if (providerNumber && isLDMKEnabled) {
@@ -156,6 +177,14 @@ function App() {
     hasSeenAnalyticsOptInPrompt,
     hasCompletedOnboarding,
   ]);
+
+  useEffect(() => {
+    if (!datadogFF?.enabled) return;
+    initializeDatadogProvider(
+      datadogFF?.params as Partial<PartialInitializationConfiguration>,
+      isTrackingEnabled ? TrackingConsent.GRANTED : TrackingConsent.NOT_GRANTED,
+    );
+  }, [datadogFF?.params, datadogFF?.enabled, isTrackingEnabled]);
 
   useAccountsWithFundsListener(accounts, updateIdentify);
   useFetchCurrencyAll();
@@ -252,12 +281,26 @@ function App() {
     lense: walletExportSelector,
   });
 
+  useDBSaveEffect({
+    save: saveLargeMoverState,
+    throttle: 500,
+    getChangesStats: (a, b) => a.largeMover !== b.largeMover,
+    lense: exportLargeMoverSelector,
+  });
+
   return (
     <GestureHandlerRootView style={styles.root}>
       <SyncNewAccounts priority={5} />
       <TransactionsAlerts />
       <ExperimentalHeader />
-      <RootNavigator />
+      {datadogFF?.enabled ? (
+        <DatadogProvider configuration={datadogAutoInstrumentation}>
+          <RootNavigator />
+        </DatadogProvider>
+      ) : (
+        <RootNavigator />
+      )}
+
       <AnalyticsConsole />
       <PerformanceConsole />
       <DebugTheme />