+++ apps/ledger-live-mobile/src/screens/Swap/History/index.tsx
@@ -3,6 +3,7 @@ import getCompleteSwapHistory from "@ledgerhq/live-common/exchange/swap/getCompl
 import { isSwapOperationPending } from "@ledgerhq/live-common/exchange/swap/index";
 import { MappedSwapOperation, SwapHistorySection } from "@ledgerhq/live-common/exchange/swap/types";
 import updateAccountSwapStatus from "@ledgerhq/live-common/exchange/swap/updateAccountSwapStatus";
+import { getParentAccount } from "@ledgerhq/live-common/account/index";
 import type { Account } from "@ledgerhq/types-live";
 import { useTheme } from "@react-navigation/native";
 import React, { useCallback, useEffect, useMemo, useRef, useState } from "react";
@@ -46,7 +47,18 @@ const History = () => {
   const syncAccounts = useSyncAllAccounts();
 
   useEffect(() => {
-    setSections(getCompleteSwapHistory(accounts));
+    const history = getCompleteSwapHistory(accounts);
+
+    // fix token account parent
+    history.forEach(section => {
+      section.data.forEach(item => {
+        if (item.toAccount.type === "TokenAccount" && !!item.toParentAccount) {
+          item.toParentAccount = getParentAccount(item.toAccount, accounts);
+        }
+      });
+    });
+
+    setSections(history);
   }, [accounts, setSections]);
 
   const refreshSwapHistory = useCallback(() => {
@@ -109,19 +121,20 @@ const History = () => {
   );
 
   const exportSwapHistory = async () => {
-    const mapped = await mappedSwapOperationsToCSV(sections);
-    const base64 = Buffer.from(mapped).toString("base64");
-    const options = {
-      title: t("transfer.swap.history.exportButton"),
-      message: t("transfer.swap.history.exportButton"),
-      failOnCancel: false,
-      saveToFiles: true,
-      type: "text/csv",
-      filename: t("transfer.swap.history.exportFilename"),
-      url: `data:text/csv;base64,${base64}`,
-    };
-
     try {
+      const mapped = mappedSwapOperationsToCSV(sections);
+
+      const base64 = Buffer.from(mapped).toString("base64");
+      const options = {
+        title: t("transfer.swap.history.exportButton"),
+        message: t("transfer.swap.history.exportButton"),
+        failOnCancel: false,
+        saveToFiles: true,
+        type: "text/csv",
+        filename: t("transfer.swap.history.exportFilename"),
+        url: `data:text/csv;base64,${base64}`,
+      };
+
       await Share.open(options);
     } catch (err) {
       // `failOnCancel: false` is not enough to prevent throwing on cancel apparently ¯\_(ツ)_/¯