+++ libs/ledgerjs/packages/types-live/src/feature.ts
@@ -151,6 +151,7 @@ export type Features = CurrencyFeatures & {
   buyDeviceFromLive: Feature_BuyDeviceFromLive;
   mockFeature: Feature_MockFeature;
   buySellUi: Feature_BuySellUiManifest;
+  buySellLoader: Feature_BuySellLoader;
   buySellShortcut: DefaultFeature;
   referralProgramDesktopSidebar: Feature_ReferralProgramDesktopSidebar;
   disableNftSend: Feature_DisableNftSend;
@@ -232,6 +233,30 @@ export type Features = CurrencyFeatures & {
   lldNetworkBasedAddAccount: DefaultFeature;
   llmOfacGeoBlocking: DefaultFeature;
   lldOfacGeoBlocking: DefaultFeature;
+  llmDatadog: {
+    enabled: boolean;
+    params: Partial<{
+      batchProcessingLevel: "MEDIUM" | "HIGH" | "LOW";
+      batchSize: "LARGE" | "MEDIUM" | "SMALL";
+      bundleLogsWithRum: boolean;
+      bundleLogsWithTraces: boolean;
+      longTaskThresholdMs: number | false;
+      nativeInteractionTracking: boolean;
+      nativeLongTaskThresholdMs: number | false;
+      nativeViewTracking: boolean;
+      resourceTracingSamplingRate: number;
+      serviceName: string;
+      sessionSamplingRate: number;
+      trackBackgroundEvents: boolean;
+      trackFrustrations: boolean;
+      trackErrors: boolean;
+      trackResources: boolean;
+      trackInteractions: boolean;
+      trackWatchdogTerminations: boolean;
+      uploadFrequency: "AVERAGE" | "FREQUENT" | "RARE";
+      vitalsUpdateFrequency: "AVERAGE" | "FREQUENT" | "RARE" | "NEVER";
+    }>;
+  };
 };
 
 /**
@@ -565,6 +590,10 @@ export type Feature_LlmRefreshMarketData = Feature<{
 
 export type Feature_BuySellUiManifest = Feature<{
   manifestId: string; // id of the app to use for the Buy/Sell UI, e.g. "buy-sell-ui"
+}>;
+
+export type Feature_BuySellLoader = Feature<{
+  durationMs: number;
 }>;
 
 export type Feature_LldWalletSync = Feature<{